finished refactoring module access
