+++ test/unit/test_pinmesh.py
@@ -2,7 +2,7 @@ import pytest
 from math import isclose
 from numpy.testing import assert_allclose
 from mpactpy.utils import ROUNDING_RELATIVE_TOLERANCE
-from mpactpy.pinmesh import RectangularPinMesh, GeneralCylindricalPinMesh
+from mpactpy import RectangularPinMesh, GeneralCylindricalPinMesh
 
 TOL = ROUNDING_RELATIVE_TOLERANCE * 1E-2
